+++ arithmetization/src/test/java/net/consensys/linea/zktracer/instructionprocessing/createTests/advanced/CustomCreate2Payload.java
@@ -66,13 +66,11 @@ public class CustomCreate2Payload {
     return Bytes.fromHexStringLenient(FunctionEncoder.encode(function));
   }
 
-  public static Bytes create2FourTimes_withRevertTrigger(boolean triggerRevert, boolean nested) {
+  public static Bytes create2FourTimes_withRevertTrigger(boolean triggerRevert) {
     Function function =
         new Function(
             CustomCreate2.FUNC_CREATE2FOURTIMES_WITHREVERTTRIGGER,
-            Arrays.asList(
-                new org.web3j.abi.datatypes.Bool(triggerRevert),
-                new org.web3j.abi.datatypes.Bool(nested)),
+            Arrays.asList(new org.web3j.abi.datatypes.Bool(triggerRevert)),
             Collections.emptyList());
     return Bytes.fromHexStringLenient(FunctionEncoder.encode(function));
   }
