--- arithmetization/src/test/java/net/consensys/linea/zktracer/instructionprocessing/createTests/advanced/InitCodeTests.java
@@ -1,835 +1,0 @@
-/*
- * Copyright Consensys Software Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-package net.consensys.linea.zktracer.instructionprocessing.createTests.advanced;
-
-import static com.google.common.base.Preconditions.checkArgument;
-import static net.consensys.linea.zktracer.Fork.isPostCancun;
-import static net.consensys.linea.zktracer.instructionprocessing.utilities.MonoOpCodeSmcs.keyPair;
-import static net.consensys.linea.zktracer.instructionprocessing.utilities.MonoOpCodeSmcs.userAccount;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-import lombok.extern.slf4j.Slf4j;
-import net.consensys.linea.UnitTestWatcher;
-import net.consensys.linea.reporting.TracerTestBase;
-import net.consensys.linea.testing.*;
-import net.consensys.linea.testing.ToyTransaction.ToyTransactionBuilder;
-import net.consensys.linea.testing.generated.ContractC;
-import net.consensys.linea.testing.generated.CustomCreate2;
-import net.consensys.linea.zktracer.instructionprocessing.utilities.SmartContractTestValidator;
-import org.apache.tuweni.bytes.Bytes;
-import org.hyperledger.besu.crypto.Hash;
-import org.hyperledger.besu.datatypes.Address;
-import org.hyperledger.besu.datatypes.Wei;
-import org.hyperledger.besu.ethereum.core.Transaction;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInfo;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.web3j.abi.EventEncoder;
-
-@Slf4j
-@ExtendWith(UnitTestWatcher.class)
-public class InitCodeTests extends TracerTestBase {
-
-  // This suite aims at testing advanced CREATE2 scenarii using smart contracts
-  // ** CustomCreate2 **
-  // CustomCreate2 is the smart contract used to pilot the deployment of a complex initCodeC
-  // CustomCreate2 stores an initCodeC and a salt used for subsequent deployments
-  // CustomCreate2 has different create2 methods to have deployment scenarii within the same
-  // transaction
-  // CustomCreate2 can CALL/STATICCALL itself and contractC
-  // ** initCodeC **
-  // initCodeC can be piloted by CustomCreate2 via the value passed in the transaction
-  // value 1 Wei enables a storage modification in ContractC
-  // value 2 Wei calls CustomCreate2 back to trigger an immediate redeployment of ContractC
-  // value 3 Wei triggers a self-destruct of ContractC
-  // value 4 Wei triggers a revert on demand
-  // ** ContractC **
-  // ContractC can modify storage, revert on demand, self-destruct on demand, and call back
-  // CustomCreate2 to trigger a redeployment of ContractC
-
-  static final Wei defaultBalance = Wei.of(4500L);
-
-  // Deployment code from testing/src/main/solidity/ContractC.sol
-  // Generated for the London EVM in testing/build/resources/main/solidity/ContractC.json
-  // Value =
-  // 6080604052600034905060003390506001820361006c578060008084815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff160217905550610196565b60028203610161578073ffffffffffffffffffffffffffffffffffffffff16630161e6d76040518163ffffffff1660e01b8152600401600060405180830381600087803b1580156100bc57600080fd5b505af19250505080156100cd575060015b61015c576100d96101cd565b806308c379a0036100fa57506100ed61026a565b806100f857506100fc565b005b505b3d8060008114610128576040519150601f19603f3d011682016040523d82523d6000602084013e61012d565b606091505b507f95a42732c1514e697b8c2992b6ee317a9bc3c4568e341cdbd58ff87fc4c467cd60405160405180910390a1005b610195565b6003820361017c5761017761019d60201b60201c565b610194565b60048203610193576101926101bb60201b60201c565b5b5b5b5b50506102fa565b60003090508073ffffffffffffffffffffffffffffffffffffffff16ff5b600080fd5b60008160e01c9050919050565b600060033d11156101ec5760046000803e6101e96000516101c0565b90505b90565b6000604051905090565b6000601f19601f8301169050919050565b7f4e487b7100000000000000000000000000000000000000000000000000000000600052604160045260246000fd5b610242826101f9565b810181811067ffffffffffffffff821117156102615761026061020a565b5b80604052505050565b600060443d106102f75761027c6101ef565b60043d036004823e80513d602482011167ffffffffffffffff821117156102a45750506102f7565b808201805167ffffffffffffffff8111156102c257505050506102f7565b80602083010160043d0385018111156102df5750505050506102f7565b6102ee82602001850186610239565b82955050505050505b90565b610379806103096000396000f3fe608060405234801561001057600080fd5b50600436106100575760003560e01c80632199ecd71461005c578063545b079a146100665780635e666e4a146100825780636c0c412e146100b2578063983d1ce2146100ce575b600080fd5b6100646100d8565b005b610080600480360381019061007b9190610249565b6100f6565b005b61009c600480360381019061009791906102ac565b610159565b6040516100a991906102e8565b60405180910390f35b6100cc60048036038101906100c79190610303565b61018c565b005b6100d66101e1565b005b60003090508073ffffffffffffffffffffffffffffffffffffffff16ff5b8073ffffffffffffffffffffffffffffffffffffffff16630161e6d76040518163ffffffff1660e01b8152600401600060405180830381600087803b15801561013e57600080fd5b505af1158015610152573d6000803e3d6000fd5b5050505050565b60006020528060005260406000206000915054906101000a900473ffffffffffffffffffffffffffffffffffffffff1681565b8060008084815260200190815260200160002060006101000a81548173ffffffffffffffffffffffffffffffffffffffff021916908373ffffffffffffffffffffffffffffffffffffffff1602179055505050565b600080fd5b600080fd5b600073ffffffffffffffffffffffffffffffffffffffff82169050919050565b6000610216826101eb565b9050919050565b6102268161020b565b811461023157600080fd5b50565b6000813590506102438161021d565b92915050565b60006020828403121561025f5761025e6101e6565b5b600061026d84828501610234565b91505092915050565b6000819050919050565b61028981610276565b811461029457600080fd5b50565b6000813590506102a681610280565b92915050565b6000602082840312156102c2576102c16101e6565b5b60006102d084828501610297565b91505092915050565b6102e28161020b565b82525050565b60006020820190506102fd60008301846102d9565b92915050565b6000806040838503121561031a576103196101e6565b5b600061032885828601610297565b925050602061033985828601610234565b915050925092905056fea26469706673582212203989edc877fe659634a18fccfea5c11d5bb1f6f616f6cacad1172e8af1c1e83064736f6c634300081a0033
-  static final String initCodeC =
-      SmartContractUtils.getSolidityContractCompiledByteCode(ContractC.class).toString();
-  static final String salt = "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef";
-  static final String buildCustomCreate2Address = "789101";
-
-  ToyAccount customCreate2Account =
-      ToyAccount.builder()
-          .address(Address.fromHexString("0x" + buildCustomCreate2Address))
-          .balance(defaultBalance)
-          .nonce(1)
-          .code(SmartContractUtils.getSolidityContractRuntimeByteCode(CustomCreate2.class))
-          .build();
-
-  static final Long gasLimit = 10000000L;
-
-  // Compute expected address for ContractC with Create2
-  // address = keccak256(0xff + sender_address + salt + keccak256(initialisation_code))[12:]
-  String senderAddNoOx = customCreate2Account.getAddress().toHexString().substring(2);
-  String saltNoOx = salt.substring(2);
-  String hashInitCodeCNo0x =
-      Hash.keccak256(Bytes.fromHexString(initCodeC)).toHexString().substring(2);
-  Bytes expectedContractCAddress =
-      Hash.keccak256(Bytes.fromHexString("0xff" + senderAddNoOx + saltNoOx + hashInitCodeCNo0x))
-          .slice(12);
-  // padding left to fit the 32 bytes log data format
-  Bytes expectedContractCAddressLogData =
-      Bytes.fromHexString(
-          "0x000000000000000000000000" + expectedContractCAddress.toString().substring(2));
-
-  // Payloads preparation
-  Bytes storeInitCodeC = CustomCreate2Payload.storeInitCodeC(initCodeC);
-  Bytes storeSalt = CustomCreate2Payload.storeSalt(salt);
-  Bytes create2WithInitCodeC = CustomCreate2Payload.create2WithInitCodeC_withValueAndRevert();
-  Bytes callContractCStoreInMapPayload =
-      CustomCreate2Payload.callContractC(
-          ContractCPayload.storeInMap(1, "0x0000000000000000000000000000000000001234"), false);
-  Bytes callContractCSelfDestructPayload =
-      CustomCreate2Payload.callContractC(ContractCPayload.selfDestructOnDemand(), false);
-  Bytes create2WithCallBackAfterCreate2_noValue =
-      CustomCreate2Payload.create2WithCallCtoCallback_noValue(false);
-  Bytes callMyselfWithCreate2WithCallCtoCallback_noValueAndNested =
-      CustomCreate2Payload.callMyself(
-          CustomCreate2Payload.create2WithCallCtoCallback_noValue(true), false, 9000000);
-  Bytes create2WithInitCodeC_withValue =
-      CustomCreate2Payload.create2WithInitCodeC_withValueAndRevert();
-
-  Bytes create2CallC_withRevert = CustomCreate2Payload.create2CallC_withRevertTrigger(true, false);
-  Bytes callMyselfWithCreate2CallC_withRevertAndNested =
-      CustomCreate2Payload.callMyself(
-          CustomCreate2Payload.create2CallC_withRevertTrigger(true, true), false, 9000000);
-  Bytes create2CallC_noRevert = CustomCreate2Payload.create2CallC_withRevertTrigger(false, false);
-  Bytes create2WithStaticCall = CustomCreate2Payload.create2WithStaticCall(false);
-  Bytes callMyselfWithCreate2WithStaticCall_nested =
-      CustomCreate2Payload.callMyself(
-          CustomCreate2Payload.create2WithStaticCall(true), false, 9000000);
-  // Bytes create2FourTimes = CustomCreate2Payload.create2FourTimes();
-  Bytes create2FourTimes_noRevert =
-      CustomCreate2Payload.create2FourTimes_withRevertTrigger(false, false);
-  Bytes create2FourTimes_withRevert =
-      CustomCreate2Payload.create2FourTimes_withRevertTrigger(true, false);
-  Bytes create2WithCallCtoCallback_noValue =
-      CustomCreate2Payload.create2WithCallCtoCallback_noValue(false);
-  Bytes create2WithCallCtoCallback_noValueAndNested =
-      CustomCreate2Payload.create2WithCallCtoCallback_noValue(true);
-
-  Bytes callCToModifyStorageAndSelfdestruct =
-      CustomCreate2Payload.callCToModifyStorageAndSelfdestruct();
-
-  // Logs for transaction validator
-  String contractCreatedEvent = EventEncoder.encode(CustomCreate2.CONTRACTCREATED_EVENT);
-  String staticCallMyselfFailEvent = EventEncoder.encode(CustomCreate2.STATICCALLMYSELFFAIL_EVENT);
-  String callCreate2WithInitCodeC_withValue_Event =
-      EventEncoder.encode(CustomCreate2.CALLCREATE2WITHINITCODEC_WITHVALUE_EVENT);
-  String callCreate2WithInitCodeC_noValue_Event =
-      EventEncoder.encode(CustomCreate2.CALLCREATE2WITHINITCODEC_NOVALUE_EVENT);
-  String callMyselfFail = EventEncoder.encode(CustomCreate2.CALLMYSELFFAIL_EVENT);
-  String callContractCFailEvent = EventEncoder.encode(CustomCreate2.CALLCONTRACTCFAIL_EVENT);
-  String storeInMapEvent = EventEncoder.encode(ContractC.STOREINMAP_EVENT);
-  String immediateRedeploymentFailEvent =
-      EventEncoder.encode(ContractC.IMMEDIATEREDEPLOYMENTFAIL_EVENT);
-  String selfDestructEvent = EventEncoder.encode(ContractC.SELFDESTRUCT_EVENT);
-
-  @Test
-  void deployContractCWithCreate2(TestInfo testInfo) {
-
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-
-    // List all logs expected from each topic
-    // In the last transaction, we attempt to redeploy contractC at the same address (see detailed
-    // description below). However, as the previous self-destruct was not done in the same tx as the
-    // 1st deployment, the re-deployment now fails.
-    int lastTxIsContractCreatedEvent = isPostCancun(fork) ? 0 : 1;
-    logsTopicMap.put(
-        contractCreatedEvent, List.of(0, 0, 1, 0, 0, 0, 0, 0, lastTxIsContractCreatedEvent));
-    logsTopicMap.put(storeInMapEvent, List.of(0, 0, 0, 1, 0, 0, 0, 0, 1));
-    logsTopicMap.put(staticCallMyselfFailEvent, List.of(0, 0, 0, 0, 0, 0, 0, 1, 0));
-    logsTopicMap.put(callCreate2WithInitCodeC_withValue_Event, List.of(0, 0, 1, 0, 0, 0, 0, 0, 0));
-    // List data expected for each topic
-    Bytes lastTxContractCreatedEvent =
-        isPostCancun(fork) ? Bytes.EMPTY : expectedContractCAddressLogData;
-    logsDataMap.put(
-        contractCreatedEvent,
-        List.of(
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            expectedContractCAddressLogData,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            lastTxContractCreatedEvent));
-    logsDataMap.put(
-        storeInMapEvent,
-        List.of(
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.fromHexString(
-                "0x0000000000000000000000000000000000000000000000000000000000000001"),
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.fromHexString(
-                "0x0000000000000000000000000000000000000000000000000000000000000000")));
-    // List status expected per transaction
-    // 0 is FAILED
-    // 1 is SUCCESSFUL
-    int lastTxStatus = isPostCancun(fork) ? 0 : 1;
-    List<Integer> txStatuses = List.of(1, 1, 1, 1, 1, 0, 0, 1, lastTxStatus);
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2Validator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            // The list below contains `Bytes` payloads
-            // Every item in that list will define a unique transaction having it as its payload.
-            List.of(
-                // CustomCreate2 initialization
-                storeInitCodeC,
-                storeSalt,
-                // SCENARIO 1 - Deploy ContractC through CustomCreate2, modify storage,
-                // self-destruct ContractC
-                // ContractC is deployed
-                // ContractC Storage is modified
-                // ContractC is self-destructed
-                // TXSTATUS : Successfull
-                // LOGS: 1 CalledCreate2WithInitCodeC + 1 ContractCreated
-                // Note : 1 CREATE2 opcode called
-                create2WithInitCodeC,
-                callContractCStoreInMapPayload,
-                callContractCSelfDestructPayload,
-                // SCENARIO 2 - Deploy ContractC and attempt redeployment after in same transaction
-                // Transaction reverts, nothing is deployed
-                // TXSTATUS : Failed
-                // LOGS: no logs
-                // Note : 2 CREATE2 opcode called
-                create2WithCallBackAfterCreate2_noValue,
-                // SCENARIO 3 - Deploy ContractC and the deployment attempts redeployment
-                // The ContractC deployment is done with value 2 - this value pilots the initcode so
-                // immediate redeployment is attempted
-                // While deploying ContractC adds STOP opcode after immediate redeployment attempt
-                // has failed
-                // ContractC is deployed with empty bytecode
-                // Call ContractC to modify storage
-                // Revert on demand
-                // Transaction reverts
-                // TXSTATUS : Failed
-                // LOGS: no logs
-                // Note : 2 CREATE2 opcode called
-                create2CallC_withRevert,
-                // SCENARIO 4 - Attempt ContractC deployment with a staticCall
-                // TXSTATUS : Successfull
-                // LOGS: 1 StaticCallMyselfFail
-                // Note 1 : no CalledCreate2WithInitCodeCEvent as attempt fails prior
-                // Note : 1 CREATE2 opcode called
-                create2WithStaticCall,
-                // SCENARIO 5 - Four ContractC deployment attempts : (1) with max value, (2)
-                // acceptable value, (3) max value and
-                // (4) acceptable value
-                // (1) is aborted
-                // (2) results in deployment of ContractC
-                // (3) is aborted
-                // (4) fails as it's a collision with attempt (2)
-                // TXSTATUS : Successfull
-                // LOGS: 1 ContractCreated
-                // Note : 4 CREATE2 opcode called
-                create2FourTimes_noRevert),
-            // Values to pilot initCode : as many as there are transactions
-            List.of(0L, 0L, 0L, 0L, 0L, 0L, 2L, 0L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2Validator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-
-    // Final check on the deployment number of ContractC
-
-    // PRE-CANCUN FORK
-    // At start, deploymentNumber = 0
-    // transaction 3 - create2WithInitCodeC : deploymentNumber ++
-    // - Deploys contract C with non empty code
-    // - deploymentNumber = 1
-    // transaction 5 - callContractCSelfDestructPayload : deploymentNumber ++
-    // - Self-destruct successful increments deployment number
-    // - deploymentNumber = 2
-    // transaction 6 - create2WithCallBackAfterCreate2 : deploymentNumber ++
-    // - First create2 is successful so increments the deployment number, second create2 makes the
-    // whole transaction revert
-    // - deploymentNumber = 3
-    // transaction 7 - create2CallCAndRevert : deploymentNumber ++
-    // - Create2 deploys contractC with empty bytecode (thus the deployment number increment) and
-    // reverts
-    // - deploymentNumber = 4
-    // transaction 9 - create2FourTimes : deploymentNumber ++
-    // - Only one create2 is successful so increments the deployment number by 1
-    // - deploymentNumber = 5
-
-    // POST-CANCUN FORK
-    // At start, deploymentNumber = 0
-    // transaction 3 - create2WithInitCodeC : deploymentNumber ++
-    // - Deploys contract C with non empty code
-    // - deploymentNumber = 1
-    int deploymentNumber =
-        toyExecutionEnvironmentV2
-            .getHub()
-            .transients()
-            .conflation()
-            .deploymentInfo()
-            .deploymentNumber(Address.fromHexString(expectedContractCAddress.toString()));
-    int expectedDeploymentNumber = isPostCancun(fork) ? 1 : 5;
-    assertEquals(expectedDeploymentNumber, deploymentNumber);
-  }
-
-  @Test
-  void deployContractCWithCreate2ScenariiTriggeredFromRoot(TestInfo testInfo) {
-    // Payload preparation
-    Bytes advancedCreateScenariiTriggeredFromRoot =
-        CustomCreate2Payload.advancedCreateScenariiTriggeredFromRoot(initCodeC, salt);
-
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-
-    // int lastTxIsContractCreatedEvent = isPostCancun(fork) ? 0 : 1;
-    // int lastTxStatus = isPostCancun(fork) ? 0 : 1;
-    List<Integer> txStatuses = List.of(1, 1);
-
-    // 1 in scenario 1
-    // 1 in scenario 3
-    logsTopicMap.put(callMyselfFail, List.of(2, 0));
-    // 1 in scenario 2
-    logsTopicMap.put(staticCallMyselfFailEvent, List.of(1, 0));
-    // 1 in scenario 4
-    logsTopicMap.put(contractCreatedEvent, List.of(1, 1));
-    logsTopicMap.put(callCreate2WithInitCodeC_noValue_Event, List.of(1, 0));
-    // 1 in scenario 5
-    logsTopicMap.put(storeInMapEvent, List.of(1, 0));
-    logsTopicMap.put(selfDestructEvent, List.of(1, 0));
-    // In the last create2 from the last transaction, to see if selfdestruct from previous
-    // transaction has worked
-    logsTopicMap.put(callCreate2WithInitCodeC_withValue_Event, List.of(0, 1));
-
-    logsDataMap.put(
-        contractCreatedEvent,
-        List.of(expectedContractCAddressLogData, expectedContractCAddressLogData));
-    // The storeInMap call uses key 1
-    logsDataMap.put(
-        storeInMapEvent,
-        List.of(
-            Bytes.fromHexString(
-                "0x0000000000000000000000000000000000000000000000000000000000000001"),
-            Bytes.EMPTY));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(advancedCreateScenariiTriggeredFromRoot, create2WithInitCodeC_withValue),
-            List.of(2L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  @Test
-  void deployContractCWithCreate2ScenariiNestedCalls(TestInfo testInfo) {
-    // Payload preparation
-    Bytes advancedCreateScenariiNestedCalls =
-        CustomCreate2Payload.advancedCreateScenariiNestedCalls(initCodeC, salt);
-
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-
-    // int lastTxIsContractCreatedEvent = isPostCancun(fork) ? 0 : 1;
-    // int lastTxStatus = isPostCancun(fork) ? 0 : 1;
-    List<Integer> txStatuses = List.of(1, 1);
-
-    // 1 in call to scenario 3
-    logsTopicMap.put(callMyselfFail, List.of(1, 0));
-    // 1 in scenario 4
-    logsTopicMap.put(contractCreatedEvent, List.of(1, 1));
-    logsTopicMap.put(callCreate2WithInitCodeC_noValue_Event, List.of(1, 0));
-    // 1 in scenario 5
-    logsTopicMap.put(storeInMapEvent, List.of(1, 0));
-    logsTopicMap.put(selfDestructEvent, List.of(1, 0));
-    // In the last create2 from the last transaction, to see if selfdestruct from previous
-    // transaction has worked
-    logsTopicMap.put(callCreate2WithInitCodeC_withValue_Event, List.of(0, 1));
-
-    logsDataMap.put(
-        contractCreatedEvent,
-        List.of(expectedContractCAddressLogData, expectedContractCAddressLogData));
-    // The storeInMap call uses key 1
-    logsDataMap.put(
-        storeInMapEvent,
-        List.of(
-            Bytes.fromHexString(
-                "0x0000000000000000000000000000000000000000000000000000000000000001"),
-            Bytes.EMPTY));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(advancedCreateScenariiNestedCalls, create2WithInitCodeC_withValue),
-            List.of(2L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 1 - NO REVERT AT THE END
-  // CREATE2 four times
-  // (1) with max value - aborted
-  // (2) acceptable value - ContractC deployed
-  // (3) max value - aborted
-  // (4) acceptable value - ContractC deployment fails as it's a collision with attempt (2)
-  // TXSTATUS : Successful
-  // LOGS: 1 ContractCreated, 1 StoreInMap
-  // Note : 4 CREATE2 opcode called
-  // Note 2 : transactions are sent with no values
-  @Test
-  void deployScenario1NoRevert(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-
-    List<Integer> txStatuses = List.of(1, 1, 1);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 1));
-    logsTopicMap.put(storeInMapEvent, List.of(0, 0, 1));
-    logsDataMap.put(
-        contractCreatedEvent, List.of(Bytes.EMPTY, Bytes.EMPTY, expectedContractCAddressLogData));
-    // The storeInMap call uses msg.value as the key, here it's 0
-    logsDataMap.put(
-        storeInMapEvent,
-        List.of(
-            Bytes.EMPTY,
-            Bytes.EMPTY,
-            Bytes.fromHexString(
-                "0x0000000000000000000000000000000000000000000000000000000000000000")));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, create2FourTimes_noRevert),
-            List.of(0L, 0L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 1 - NO REVERT AT THE END
-  // CREATE2 four times
-  // (1) with max value - aborted
-  // (2) acceptable value - ContractC deployed
-  // (3) max value - aborted
-  // (4) acceptable value - ContractC deployment fails as it's a collision with attempt (2)
-  // Revert on demand
-  // TXSTATUS : Failed
-  // LOGS: 0 ContractCreated, 0 StoreInMap
-  // Note : 4 CREATE2 opcode called
-  // Note 2 : transactions are sent with no values
-  @Test
-  void deployScenario1(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-
-    List<Integer> txStatuses = List.of(1, 1, 0);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 0));
-    logsTopicMap.put(storeInMapEvent, List.of(0, 0, 0));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, create2FourTimes_withRevert),
-            List.of(0L, 0L, 2L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 2 - STATIC CALL A CREATE2
-  // Attempt to static call a create2 deployment
-  // TXSTATUS : Successful
-  // LOGS: 1 StaticCallMyselfFail
-  // Note: transaction is sent with no values
-  @Test
-  void deployScenario2(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1);
-
-    logsTopicMap.put(staticCallMyselfFailEvent, List.of(1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, new HashMap<>());
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account, userAccount, List.of(create2WithStaticCall), List.of(0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 2 - STATIC CALL A CREATE2, WITH SCENARIO 1 NESTED
-  // Attempt to static call a create2 deployment, after create2 four times call
-  // TXSTATUS : Successful
-  // LOGS: 1 CallMyselfFail, 1 StaticCallMyselfFail
-  // Note: transaction is sent with no values
-  @Test
-  void deployScenario2Nested(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1);
-
-    // 1 from scenario 1 nested
-    logsTopicMap.put(callMyselfFail, List.of(1));
-    // 1 from scenario 2 static call
-    logsTopicMap.put(staticCallMyselfFailEvent, List.of(1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, new HashMap<>());
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(callMyselfWithCreate2WithStaticCall_nested),
-            List.of(0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 3 - ATTEMPT CREATE2 WITHIN A CREATE2, NO REVERT AT THE END
-  // Attempts a create 2 within a create2, goes back to initCodeC and stops
-  // Code deployed is empty. We call contract C to modify storage
-  // TXSTATUS : Successful
-  // LOGS: 1 ContractCreated, 1 ImmediateRedeploymentFail, 0 CallCreate2WithInitCodeC_withValue
-  // because the create2 reverts, 0 StoreInMap as the code is empty
-  // Note: transaction is sent with value 2 to do a create2 within a create2
-  @Test
-  void deployScenario3NoRevert(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 1);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 1));
-    // Callback to attempt a create2 within a create2, triggered by the msg.value == 2
-    logsTopicMap.put(immediateRedeploymentFailEvent, List.of(0, 0, 1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, create2CallC_noRevert),
-            List.of(0L, 0L, 2L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 3 - ATTEMPT CREATE2 WITHIN A CREATE2, WITH REVERT AT THE END
-  // Attempts a create 2 within a create2, goes back to initCodeC and stops
-  // Code deployed is empty. We call contract C to modify storage. We revert the whole transaction
-  // TXSTATUS : Failed
-  // LOGS: 0 ContractCreated, O ImmediateRedeploymentFail, 0 CallCreate2WithInitCodeC_withValue
-  // because the create2 reverts, 0 StoreInMap as the code is empty
-  // Note: transaction is sent with value 2 to do a create2 within a create2
-  @Test
-  void deployScenario3(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 0);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 0));
-    // Callback to attempt a create2 within a create2, triggered by the msg.value == 2
-    logsTopicMap.put(immediateRedeploymentFailEvent, List.of(0, 0, 0));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, create2CallC_withRevert),
-            List.of(0L, 0L, 2L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 3 - ATTEMPT CREATE2 WITHIN A CREATE2, WITH REVERT AT THE END, WITH SCENARIO 1 AND 2
-  // NESTED
-  // Attempts a create 2 within a create2, goes back to initCodeC and stops
-  // Code deployed is empty. We call contract C to modify storage. We revert the whole transaction
-  // TXSTATUS : Failed
-  // LOGS: 0 ContractCreated, O ImmediateRedeploymentFail, 0 CallCreate2WithInitCodeC_withValue
-  // because the create2 reverts, 0 StoreInMap as the code is empty
-  // Note: transaction is sent with value 2 to do a create2 within a create2
-  @Test
-  void deployScenario3Nested(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 1);
-
-    logsTopicMap.put(callMyselfFail, List.of(0, 0, 1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, callMyselfWithCreate2CallC_withRevertAndNested),
-            List.of(0L, 0L, 2L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 4 - ATTEMPT CREATE2 AFTER A CREATE2
-  // Attempts a create 2 after a successful create2, by calling ContractC that does a callback to
-  // CustomCreate2
-  // TXSTATUS : Successful
-  // LOGS: 1 ContractCreated, 1 CallCreate2WithInitCodeC_noValue
-  // Note: transaction is sent with value 0
-  @Test
-  void deployScenario4(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 1);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 1));
-    // Callback to attempt a create2 within a create2, triggered by the msg.value == 2
-    logsTopicMap.put(callCreate2WithInitCodeC_noValue_Event, List.of(0, 0, 1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(storeInitCodeC, storeSalt, create2WithCallCtoCallback_noValue),
-            List.of(0L, 0L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 4 - ATTEMPT CREATE2 AFTER A CREATE2, WITH SCENARIO 1 AND 2 AND 3 NESTED
-  // Attempts a create 2 after a successful create2, by calling ContractC that does a callback to
-  // CustomCreate2
-  // TXSTATUS : Successful
-  // LOGS: 1 ContractCreated, 1 CallCreate2WithInitCodeC_noValue
-  // Note: transaction is sent with value 0
-  @Test
-  void deployScenario4Nested(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 1);
-
-    logsTopicMap.put(callMyselfFail, List.of(0, 0, 1));
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 1));
-    // Callback to attempt a create2 within a create2, triggered by the msg.value == 2
-    logsTopicMap.put(callCreate2WithInitCodeC_noValue_Event, List.of(0, 0, 1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(
-                storeInitCodeC,
-                storeSalt,
-                callMyselfWithCreate2WithCallCtoCallback_noValueAndNested),
-            List.of(0L, 0L, 2L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  // SCENARIO 5 - CREATE2, MODIFY STORAGE, SELFDESTRUCT
-  // After a successful create 2, we modify the contract storage and self-destruct
-  // TXSTATUS : Successful
-  // LOGS: 1 ContractCreated, 1 CallCreate2WithInitCodeC_noValue, 1 StoreInMap, 1 SelfDestruct
-  // Note: transaction is sent with value 0
-  @Test
-  void deployScenario5(TestInfo testInfo) {
-    Map<String, List<Integer>> logsTopicMap = new HashMap<>();
-    Map<String, List<Bytes>> logsDataMap = new HashMap<>();
-    List<Integer> txStatuses = List.of(1, 1, 1, 1);
-
-    logsTopicMap.put(contractCreatedEvent, List.of(0, 0, 1, 0));
-    logsTopicMap.put(callCreate2WithInitCodeC_noValue_Event, List.of(0, 0, 1, 0));
-    logsTopicMap.put(storeInMapEvent, List.of(0, 0, 0, 1));
-    logsTopicMap.put(selfDestructEvent, List.of(0, 0, 0, 1));
-
-    // Instantiate validator
-    TransactionProcessingResultValidator create2OneTxValidator =
-        new SmartContractTestValidator(txStatuses, logsTopicMap, logsDataMap);
-
-    List<Transaction> transactions =
-        getTransactions(
-            customCreate2Account,
-            userAccount,
-            List.of(
-                storeInitCodeC,
-                storeSalt,
-                create2WithCallCtoCallback_noValue,
-                callCToModifyStorageAndSelfdestruct),
-            List.of(0L, 0L, 0L, 0L));
-
-    final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
-        ToyExecutionEnvironmentV2.builder(chainConfig, testInfo)
-            .accounts(List.of(userAccount, customCreate2Account))
-            .transactions(transactions)
-            .transactionProcessingResultValidator(create2OneTxValidator)
-            .build();
-    toyExecutionEnvironmentV2.run();
-  }
-
-  /// /////////////////////////////////////////////////////////////////////////////////////////////
-  /// Common helpers
-  /// Create transactions with payloads/values for the same user and the same to account
-  public static List<Transaction> getTransactions(
-      ToyAccount to, ToyAccount userAccount, List<Bytes> payloads, List<Long> values) {
-
-    checkArgument(payloads.size() == values.size());
-    final List<ToyTransactionBuilder> builders = new ArrayList<>();
-
-    for (int i = 0; i < payloads.size(); i++) {
-      ToyTransactionBuilder builder =
-          ToyTransaction.builder()
-              .to(to)
-              .payload(payloads.get(i))
-              .keyPair(keyPair)
-              .gasLimit(gasLimit)
-              .value(Wei.of(values.get(i)));
-      builders.add(builder);
-    }
-    return ToyMultiTransaction.builder().build(builders, userAccount);
-  }
-}
