+++ arithmetization/src/test/java/net/consensys/linea/zktracer/instructionprocessing/createTests/advanced/StorageNoOpTests.java
@@ -14,6 +14,7 @@
  */
 package net.consensys.linea.zktracer.instructionprocessing.createTests.advanced;
 
+import static net.consensys.linea.zktracer.instructionprocessing.createTests.advanced.ScenarioUtils.getTransactions;
 import static net.consensys.linea.zktracer.instructionprocessing.utilities.MonoOpCodeSmcs.userAccount;
 
 import java.math.BigInteger;
@@ -126,7 +127,7 @@ public class StorageNoOpTests extends TracerTestBase {
     // for some reason the solidity code does not allow to pass a nonzero value to the
     // transaction; we thus provide zero Wei to both transactions
     List<Transaction> transactions =
-        InitCodeTests.getTransactions(
+        getTransactions(
             factorySmc, userAccount, List.of(deployPayload, callMainMethod), List.of(0L, 0L));
 
     final ToyExecutionEnvironmentV2 toyExecutionEnvironmentV2 =
